[I-TEST] HIV Screening (#81)
* Add new enum, `InfectionType`, to account for different viruses.

By default, `InfectionType` will be set to `HCV` to prevent breaking existing,
HCV-centric functionality, but with the addition of HIV to the model this keeps
the option of adding yet more infections easy.

* Apply the same kind of adjustments done to Linking to Screening.

Allows multiple infection types to reuse the same objects for tracking screening
quantities, specifically timing of screening and test quantity.

* Pushing the latest changes for HIV Screening, though still WIP.

Making a lot of use of enums in `HIVScreening` for the branching of background
vs intervention screening and trying to simplify the control flow a bit from
HCV screening, with plans to revisit the original later. Currently working on
making sure the Person object has the necessary methods while also aiming to cut
down on redundant methods with HIV.

* Implement `HIVScreening` event.

* Fix formatting in `build.sh`.

* Add `HIVScreeningTest.cpp` and add the first (passing) test.

* Add more HIV Screening tests.

* Remove redundant function `IsIdentifiedAsHIVInfected()`.

Because the new `screeningDetails` unordered map tracks identification on a per-
infection basis, this method and related variable are no longer necessary to be
tracked in `person::Health`.

* Standardize naming convention of `static const` members.

* Address PR comments.

* Address PR comments.

* Make HCV events explicit